+++ unit-tests/state/classic-gateways.test.js
@@ -388,6 +388,64 @@ describe("classic gateways", () => {
     beforeEach(() => {
       craig = newState();
     });
+    describe("classic_gateways.memory", () => {
+      it("should return true if a gw has memory that is not whole number", () => {
+        assert.isTrue(
+          craig.classic_gateways.memory.invalid(
+            {
+              name: "aa",
+              domain: "oops.com",
+              datacenter: "dal10",
+              network_speed: "1000",
+              public_bandwidth: "64",
+              package_key_name: "VIRTUAL_ROUTER_APPLIANCE_1_GPBS",
+              os_key_name: "OS_JUNIPER_VSRX_19_4_UP_TO_1GBPS_STANDARD_SRIOV",
+              process_key_name: "INTEL_XEON_4210_2_20",
+              private_vlan: "vlan",
+              public_vlan: "pvlan",
+              ssh_key: "key",
+              disk_key_names: ["disk"],
+              memory: -1.25,
+            },
+            {
+              craig: craig,
+              data: {
+                name: "frog",
+              },
+            }
+          ),
+          "it should be true"
+        );
+      });
+      it("should return true if a gw has memory that is whole number not in range", () => {
+        assert.isTrue(
+          craig.classic_gateways.memory.invalid(
+            {
+              name: "aa",
+              domain: "oops.com",
+              datacenter: "dal10",
+              network_speed: "1000",
+              public_bandwidth: "64",
+              package_key_name: "VIRTUAL_ROUTER_APPLIANCE_1_GPBS",
+              os_key_name: "OS_JUNIPER_VSRX_19_4_UP_TO_1GBPS_STANDARD_SRIOV",
+              process_key_name: "INTEL_XEON_4210_2_20",
+              private_vlan: "vlan",
+              public_vlan: "pvlan",
+              ssh_key: "key",
+              disk_key_names: ["disk"],
+              memory: -1,
+            },
+            {
+              craig: craig,
+              data: {
+                name: "frog",
+              },
+            }
+          ),
+          "it should be true"
+        );
+      });
+    });
     describe("classic_gateways.name", () => {
       it("should return correct helper text", () => {
         assert.deepEqual(
@@ -406,6 +464,35 @@ describe("classic gateways", () => {
       it("should return false when domain is undefined", () => {
         assert.isTrue(
           craig.classic_gateways.domain.invalid({}),
+          "it should be true"
+        );
+      });
+    });
+    describe("classic_gateways.disk_key_names", () => {
+      it("should return true if a gw has no disk key names", () => {
+        assert.isTrue(
+          craig.classic_gateways.disk_key_names.invalid(
+            {
+              name: "aa",
+              domain: "oops.com",
+              datacenter: "dal10",
+              network_speed: "1000",
+              public_bandwidth: "64",
+              package_key_name: "VIRTUAL_ROUTER_APPLIANCE_1_GPBS",
+              os_key_name: "OS_JUNIPER_VSRX_19_4_UP_TO_1GBPS_STANDARD_SRIOV",
+              process_key_name: "INTEL_XEON_4210_2_20",
+              private_vlan: "vlan",
+              public_vlan: "pvlan",
+              ssh_key: "key",
+              disk_key_names: [],
+            },
+            {
+              craig: craig,
+              data: {
+                name: "frog",
+              },
+            }
+          ),
           "it should be true"
         );
       });
@@ -451,6 +538,64 @@ describe("classic gateways", () => {
       });
     });
     describe("classic_gateways.public_network_only", () => {
+      it("should return false if a gw has no public_vlan and private_network_only is true", () => {
+        assert.isFalse(
+          craig.classic_gateways.shouldDisableSave(
+            {
+              name: "aa",
+              domain: "oops.com",
+              datacenter: "dal10",
+              network_speed: "1000",
+              public_bandwidth: "64",
+              package_key_name: "VIRTUAL_ROUTER_APPLIANCE_1_GPBS",
+              os_key_name: "OS_JUNIPER_VSRX_19_4_UP_TO_1GBPS_STANDARD_SRIOV",
+              process_key_name: "INTEL_XEON_4210_2_20",
+              private_vlan: "vlan",
+              public_vlan: "",
+              private_network_only: true,
+              ssh_key: "key",
+              disk_key_names: ["HARD_DRIVE_2_00_TB_SATA_2"],
+              memory: 64,
+            },
+            {
+              craig: craig,
+              data: {
+                name: "frog",
+              },
+            }
+          ),
+          "it should be true"
+        );
+      });
+      it("should return true if a gw has no public_vlan and private_network_only is false", () => {
+        assert.isTrue(
+          craig.classic_gateways.shouldDisableSave(
+            {
+              name: "aa",
+              domain: "oops.com",
+              datacenter: "dal10",
+              network_speed: "1000",
+              public_bandwidth: "64",
+              package_key_name: "VIRTUAL_ROUTER_APPLIANCE_1_GPBS",
+              os_key_name: "OS_JUNIPER_VSRX_19_4_UP_TO_1GBPS_STANDARD_SRIOV",
+              process_key_name: "INTEL_XEON_4210_2_20",
+              private_vlan: "vlan",
+              public_vlan: "",
+              private_network_only: false,
+              ssh_key: "key",
+              disk_key_names: ["HARD_DRIVE_2_00_TB_SATA_2"],
+              memory: 64,
+            },
+            {
+              craig: craig,
+              data: {
+                name: "frog",
+              },
+            }
+          ),
+          "it should be true"
+        );
+      });
       it("should change state data when changing from false to true", () => {
         let data = {
           private_network_only: false,